# add bench for reading & optimise