--- a/Foundation/String/Read.hs
+++ b/Foundation/String/Read.hs
@@ -1,5 +1,6 @@
 module Foundation.String.Read
     ( readInteger
+    , readIntegral
     , readNatural
     , readDouble
     , readFloatingExact