--- a/Foundation/String/Read.hs
+++ b/Foundation/String/Read.hs
@@ -3,6 +3,7 @@ module Foundation.String.Read
     , readIntegral
     , readNatural
     , readDouble
+    , readRational
     , readFloatingExact
     ) where