--- a/benchs/Fake/ByteString.hs
+++ b/benchs/Fake/ByteString.hs
@@ -11,7 +11,7 @@ module Fake.ByteString
     , readInteger
     ) where
 
-import Prelude (undefined)
+import Prelude (undefined, Maybe(..))
 
 data ByteString = ByteString
 
@@ -22,5 +22,8 @@ take        = undefined
 break   _ _ = (undefined, undefined)
 reverse     = undefined
 filter _    = undefined
+
+readInt :: ByteString -> Maybe (a,b)
 readInt _    = undefined
+readInteger :: ByteString -> Maybe (a,b)
 readInteger _ = undefined