--- a/benchs/Fake/Text.hs
+++ b/benchs/Fake/Text.hs
@@ -11,7 +11,7 @@ module Fake.Text
     , double
     ) where
 
-import Prelude (undefined)
+import Prelude (undefined, Either(..))
 
 data Text = Text
 
@@ -22,5 +22,9 @@ take        = undefined
 filter _    = undefined
 reverse     = undefined
 any         = undefined
-decimal     = undefined
-double      = undefined
+
+decimal :: Text -> Either a (b, c)
+decimal = undefined
+
+double :: Text -> Either a (b, c)
+double = undefined