After correcting the initial wrong benchmark, we're slighlty slower in all cases so far (not unexpected):

Integer:

```
benchmarking types/String/Read/Integer/10000/String
time                 200.1 ns   (197.5 ns .. 202.9 ns)
benchmarking types/String/Read/Integer/10000/Text
time                 126.0 ns   (124.1 ns .. 128.4 ns)

benchmarking types/String/Read/Integer/1234567891234567890/String
time                 798.4 ns   (786.5 ns .. 809.1 ns)
benchmarking types/String/Read/Integer/1234567891234567890/Text
time                 505.7 ns   (497.8 ns .. 513.1 ns)
```

Double:

```
benchmarking types/String/Read/Double/100.56e23/String
time                 590.6 ns   (584.3 ns .. 597.0 ns)
benchmarking types/String/Read/Double/100.56e23/Text
time                 323.7 ns   (318.4 ns .. 329.3 ns)
benchmarking types/String/Read/Double/-123.1247/String
time                 597.0 ns   (589.5 ns .. 603.8 ns)
benchmarking types/String/Read/Double/-123.1247/Text
time                 315.6 ns   (308.9 ns .. 321.6 ns)
```