@vincenthz 2017-04-17
Integer:
```
benchmarking types/String/Read/Integer/10000/String
time                 116.5 ns   (115.2 ns .. 118.0 ns)

benchmarking types/String/Read/Integer/10000/Text
time                 115.6 ns   (114.4 ns .. 117.2 ns)

benchmarking types/String/Read/Integer/1234567891234567890/String
time                 462.6 ns   (453.5 ns .. 471.1 ns)

benchmarking types/String/Read/Integer/1234567891234567890/Text
time                 489.0 ns   (479.0 ns .. 496.2 ns)
```

Int:

```
benchmarking types/String/Read/Int/12345/String
time                 34.93 ns   (34.31 ns .. 35.61 ns)

benchmarking types/String/Read/Int/12345/Text
time                 36.91 ns   (36.31 ns .. 37.51 ns)

benchmarking types/String/Read/Int/12345/ByteString
time                 28.36 ns   (27.71 ns .. 29.15 ns)
```