@vincenthz 2017-04-18
55f655b96986a46d822185f946caf45596178c59 make readDouble less imprecise and fix #243