@vincenthz 2017-04-18
So now, foundation is more precise and faster than Text to parse Double.

```
benchmarking types/String/Read/Double/100.56e23/String
time                 246.4 ns   (242.7 ns .. 250.6 ns)
             
benchmarking types/String/Read/Double/100.56e23/Text
time                 315.3 ns   (310.2 ns .. 319.9 ns)
             
benchmarking types/String/Read/Double/-123.1247/String
time                 223.4 ns   (220.1 ns .. 226.2 ns)
             
benchmarking types/String/Read/Double/-123.1247/Text
time                 308.8 ns   (304.1 ns .. 313.1 ns)
```