@vincenthz 2017-04-18
For the precision thing, this is always going to be a tradeoff. It would be more precise to work out the encoding using `fromRat''` without a `/` operation, but it would also be very likely slower (and also approximate nonetheless). In practice, I haven't work out how much imprecision that add, but that would be surprising it that was that significative.

What about adding a readRational that keeps the precision, and then the user can manually call `fromRational` which would brings the same algorithm used by `read @ Double` ?