@ndmitchell 2017-04-18
readRational seems like an excellent idea.