@vincenthz 2017-04-21
it's all in now. probably could do more optimisation, but that's good enough for this iteration I think.